qwipo-ml-engine-starter/requirements.txt
fastapi
uvicorn
pandas
numpy
scipy
scikit-learn
sentence-transformers
python-dotenv
requests
jupyter
